build: use .opt variant of ocamldoc if available.
The change to configure.in is fairly simple: look for ocamldoc.opt, if it's
not available, try to use ocamldoc.
Unlike for the compilers (ocamlc and ocamlopt), we don't check that the
versions of the two match (and if they don't, use the bytecode version).
I believe it wasn't necessary to make the code more complex since an old
ocamldoc has way fewer bad consequences than a buggy compiler.

That said, we should migrate to configure.ac one day.

Btw, on my machine, "make doc" is almost 4 times faster with ocamldoc.opt.