README: revamp: reorganize sections, change titles, use asciidoc syntax.
After the previous commit that didn't change the content of the README
but simply switched it to basic asciidoc syntax, this commit has many
more changes.

Among others, I tried to change the file to make it better for a
beginner: move the sections that are the most important to them to the
top of the file.

I've also reworded some sentences and titles along with many more small
changes. I've tried to keep the content the same, only change the way it
was said and displayed.