[build] Add support for building with Dune.
We add preliminary support for building with Dune to lablgtk2; this
seems to work well and does provide some advantages:

- compositional build with other developments, for example CoqIDE,
- it allows for an easier layout / library reorganization if wished,
- automatic management of install targets, opam install file, META, stubs,
- automatic generation of documentation with `odoc`.

I have used a more "explicit" dune-file style for now, which while a
bit more verbose, should be easier to understand.

While the PR is preliminary, it has been tested with CoqIDE and works
fine.

One strong motivation to do this PR was to help the GTK 3 porting
effort and this provides a seamless developer experience.

Deferred to subsequent PRs:

- linting? The large amount of warning emitted by the Dune build
  should be addressed,
- OSX, win32 support,
- make all examples work again,
- cleanup makefiles, remove old make system?
- generate verbose rules automatically [may not be worth it]