lablgladecc ported to GBuilder and new XML DTD
Right now the output of lablgladecc is not backward
compatible with that of the old version based on
libglade. In particular, the old lablgladecc built
the windows on-demand, while the new one builds each
window exactly once at module execution time.

I will try next to restore the old behaviour.