Disappeared widgets removed
"lablgladecc -test" works again

NOTE: line 395 is a terrible hack to avoid generating
polymorphic classes that depends on an arbitrary number
of parameters. The hack was already there in lablgtk2, but
for the GList.clist class. Because of the hack, the generated
code is NOT polymorphic enough and the user has to patch the
generated code to write down the wanted type, which is BAD.