Code strengthening
I kept the code logic that came from lablgtk2 lablgladecc.ml code.
However, either that code assumed a DTD that changed a lot, or it
was fragile. This commit tries to make the code more robust by
detecting a lot of cases where (my) assumptions on the DTD could be
violated, failing gracefully (via assert false). If an assert false
will be triggered, it means that the code logic needs to be
redesigned from scratch to allow for arbitrary well-formed XML
files.

Note: the XML file format is "open", because objects that implement
a certain interface can augment the (implicit) DTD in arbitrary
ways. The main assumption in the current code is that new XML
elements are not allowed to contain recursively <object>s.