lablgladecc: first working version (#27)
* glade lib is no more, but lablgladecc can be ported
* lablgladecc ported to GBuilder and new XML DTD
* new methods bound: new and add_from_file
* gtk_builder_add_objects_from_file bound
* old behavior restored: load objects on demand
* Disappeared widgets removed
"lablgladecc -test" works again

NOTE: line 395 is a terrible hack to avoid generating
polymorphic classes that depends on an arbitrary number
of parameters. The hack was already there in lablgtk2, but
for the GList.clist class. Because of the hack, the generated
code is NOT polymorphic enough and the user has to patch the
generated code to write down the wanted type, which is BAD.

* safe dynamic casts restored in lablgladecc output

gBuilder methods now return a "unit obj" in place of
an "'a obj", like lablglade used to do. That forces
the caller to use a "Module.cast" to safely (but
dynamically) cast from "unit obj" to "something obj"

* Fix typo that prevented subtyping between tree_*
* GPack.size_group implemented
* List of recognizable objects updated and fixed
Note: this was such a painful experience. It could be
automated if only the binding of widgets was more uniform.
For example, for some widgets there is a GtkXXProps.YY module
but no corresponding GtkXX.YY module, etc.

* Options -trace -check-all removed
-check_all was no longer working in lablgtk2 as well
-trace cannot be reproduced using gBuilder

* Allow to use ~translation_domain for each widget
* lablgladecc -embed restored via GBuilder.*_from_string
- missing *_from_string methods bound
- lablgladecc -embed functionality restored

* Code strengthening
I kept the code logic that came from lablgtk2 lablgladecc.ml code.
However, either that code assumed a DTD that changed a lot, or it
was fragile. This commit tries to make the code more robust by
detecting a lot of cases where (my) assumptions on the DTD could be
violated, failing gracefully (via assert false). If an assert false
will be triggered, it means that the code logic needs to be
redesigned from scratch to allow for arbitrary well-formed XML
files.

Note: the XML file format is "open", because objects that implement
a certain interface can augment the (implicit) DTD in arbitrary
ways. The main assumption in the current code is that new XML
elements are not allowed to contain recursively <object>s.