README and examples/glade ported to lablgladecc3
Note: I did not bind the automatic management of signals
that are specified in the XML file produced by Glade. I could,
but since the callbacks can only be C code, it is not very useful.
The examples, however, used to automatically connect the quit
button to gtk_main_quit. I have preferred to add the code to the
OCaml side and call GMain.quit instead, which seems safer anyway.