[build] Teach dune about generation from ml4 files.
We do a warnings cleanup on the way.

The camlp4 rule has to be a bit special to ensure the loading of
`pr_o` as for promotion to work better; by default, `camlp4o` will use
Camlp4AutoPrinter which does:

```ocaml
if Unix.isatty Unix.stdout then
  enable_ocaml_printer ()
else
  enable_dump_ocaml_ast_printer ();
```

which turns out to be very inconvenient.