[opam] [dune-release] Fix opam metadata to reflect recent release.
After 3.0.beta4, the loose constraints should mostly work; depending
on the changes we do of course.

We also update the `homepage` field of the OPAM files, and the CHANGES
file, if I understand correctly this should indeed make `dune-release`
work out of the box [and to produce sensible changelogs]

Note that dune-release was incorrectly reading `CHANGES.API` to create
the OPAM changelog! So I had to fix that too.