[build] Use newer configurator API, correct build-deps
We use the new configurator API which does provide better error
messages and avoids some warnings during build. This means we have to
bump to dune 1.8.0 but that should not be a problem these days.

We also correct the opam files as to remove the `build` dune flag [as
per upstream policy], and require a strict matching on the version
field for depending packages [safer and few downsides]