Replace Camlp5 stream parsing with PMS
Where PMS stands for Poor-Man Streams, not what you may have thought.

What i did is that I've added a lexer,  but I'm accessing it directly,
but through a wrapper struct
```OCaml
type t = {
  mutable lookahead: token option;
  lexbuf: Lexing.lexbuf;
}```

This gives us the familar stream operations,
`next`, and `peek` and that is enough to
port the stream parsing code almost verbatim.